requirements.txt
aiofiles==23.2.1
aiosqlite==0.17.0
annotated-types==0.6.0
click==8.1.7
distlib==0.3.7
filelock==3.13.0
h11==0.14.0
html5tagger==1.3.0
httptools==0.6.1
iso8601==1.1.0
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
platformdirs==3.11.0
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
pypika-tortoise==0.1.6
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
ruff==0.1.9
sanic==23.12.0
sanic-ext==23.12.0
sanic-routing==23.12.0
setuptools==68.2.2
tenacity==8.2.3
tortoise-orm==0.20.0
tracerite==1.1.1
typing_extensions==4.9.0
ujson==5.9.0
uvloop==0.19.0
virtualenv==20.25.0
websockets==12.0
WTForms==3.1.1
